+++ src/components/Button/Button.types.ts
@@ -1,7 +1,15 @@
 import type { ButtonProps as AriaButtonProps } from 'react-aria-components';
 
+export type ButtonVariant =
+  | 'solid'
+  | 'outline'
+  | 'ghost'
+  | 'primary'
+  | 'secondary'
+  | 'tertiary';
+
 export type ButtonProps = AriaButtonProps & {
   isLoading?: boolean;
-  variant?: 'solid' | 'outline' | 'ghost';
+  variant?: ButtonVariant;
   size?: 'sm' | 'md' | 'lg';
 };