+++ src/components/LinkAsButton/LinkAsButton.types.ts
@@ -1,0 +1,7 @@
+import type { LinkProps } from '../Link/Link.types';
+import type { ButtonProps } from '../Button/Button.types';
+
+export type LinkAsButtonProps = LinkProps &
+  Pick<ButtonProps, 'variant' | 'size'> & {
+    isLoading?: boolean;
+  };