+++ src/components/LinkAsButton/index.ts
@@ -1,0 +1,2 @@
+export { LinkAsButton } from './LinkAsButton';
+export type { LinkAsButtonProps } from './LinkAsButton.types';